Create Shapes_Edited
I changed the sizing and placement of the original SVG shapes, as well as changing the color scheme to make each shape have some shade of gold or yellow. The gradients and opacities changed, as I was trying to overlap colors but keep visually distinguishable gradient lines, which wasn't completely successful. I also created a new transformation for the circle to follow.